% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_build.R
\name{shield_build}
\alias{shield_build}
\title{Build a png shield}
\usage{
shield_build(
  stub,
  label,
  logo,
  color = "blue",
  filename = "generated_badge.png",
  method = "shields.io",
  verbose = FALSE
)
}
\arguments{
\item{stub}{The left hand side of the shield}

\item{label}{The right hand side of the shield}

\item{logo}{Either the path to a png logo (ideal is 14x14 pixels) or a
pre-encoded base64 image (keep it small as this passed as a call!).}

\item{color}{What color you want the badge to be.}

\item{filename}{Where to output the file - can be a path, but must end
in png.}

\item{method}{Which service should be used for rendering. Right now it's
just shields.io}

\item{verbose}{Whether to print messages or not}
}
\description{
This function will create a local png shield
}
