% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_logo_encode.R
\name{shield_logo_encode}
\alias{shield_logo_encode}
\title{Convert a png into base64 string}
\usage{
shield_logo_encode(logo)
}
\arguments{
\item{logo}{The left hand side of the shield}
}
\description{
This function will create a local png shield
}
